// src/server/model/eventItemBasic.ts
/**
 * Generated by orval v6.11.0 🍺
 * Do not edit manually.
 * Maintenance Insights Service
 * Maintenance Insighs Service provides analytical insights about maintenance events
 * OpenAPI spec version: 1.0.0
 */
import type { EventAttributes } from './eventAttributes';
import type { EventType } from './eventType';

/**
 * Event object
 */
export interface EventItemBasic {
	attributes: EventAttributes;
	/** Optional link when comparing events */
	comparison?: string;
	self: string;
	title?: string;
	type: EventType;
}